% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndSimpleExample-cnd-conditions.R
\name{cndSimpleExample-cnd-conditions}
\alias{cndSimpleExample-cnd-conditions}
\alias{cndSimpleExample:bad_argument}
\alias{bad_argument}
\alias{cndSimpleExample:bad_argument/error}
\title{Conditions for \code{cndSimpleExample}}
\description{
Conditions for \code{cndSimpleExample}
}
\details{
The following conditions are defined in the \code{{cndSimpleExample}} package.
}
\section{\code{\link[=cnd-package]{\{cnd\}}}}{

These conditions are made with the \code{{cnd}} package though the use of
\code{\link[cnd:condition]{cnd::condition()}}.
}

\section{\code{{cndSimpleExample}} conditions}{


\subsection{\code{cndSimpleExample:bad_argument/error}}{
\describe{
\item{package}{\code{{cndSimpleExample}}}
\item{class}{\code{cndSimpleExample:bad_argument}}
\item{type}{\strong{error}}
}
x should be \code{TRUE}
}
}

\seealso{
\link[cnd:cnd-package]{cnd::cnd-package} \link[cnd:condition]{cnd::condition}
}
\keyword{internal}
